feat: Add usePlanStatus hook
