+++ frontend/feature/demand-planning/lib/index.ts
@@ -1,23 +1,3 @@
 import getApiClient from '../../../lib/api';
 
 export const apiClient = getApiClient('demand-planning');
-
-export type PlanStatusStyle = {
-    label: string;
-    color: string;
-};
-
-export const planStatusStyles: { [status: string]: PlanStatusStyle } = {
-    draft: {
-        label: 'Draft',
-        color: 'teal.300',
-    },
-    forecasted: {
-        label: 'Forecasted',
-        color: 'blue.300',
-    },
-    reviewed: {
-        label: 'Reviewed',
-        color: 'purple.300',
-    },
-};
